Backend service RPCs (#24)
This PR defines a new gRPC service for a remote backend implementation for the Durable Task Framework (DTFx). It depends on many of the existing message types from orchestration_service.proto, such as event history definitions, but defines a few new message types as well.

Implementations of this RPC could be used to build client and server components for custom DTFx backends.